Refactor Home repository to remove Android dependencies
